stack program
Stack is a Last In First Out (LIFO) data structure. Elements can Add in the top of Stack and also Deleted from top of Stack. A real-life example of stack is "arranging disk in holder". In Stack their are some functions are required to perform operation like isFull() and isEmpty() :
1. Push: use for inserting the single item at a time in stack
2. Pop: uses for deleting a single item at a time from stack
3. Traverse: to display all the elements in stack